requirements.txt
python-telegram-bot==20.7
fastapi>=0.110
uvicorn>=0.27
asyncpg>=0.29
